pyproject.toml
[project]
name = "memory-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.10.0",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
]
